fix: devolve menu antigo e move arquivos de cabeçalhos para a pasta components
