+++ smartmenu.client/src/App.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Header from './Header';
-import Menu from './Menu';
-import Footer from './Footer';
+import Header from './Components/Header';
+import Menu from './Components/Menu';
+import Footer from './Components/Footer';
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 function App() {
