+++ smartmenu.client/src/Components/Footer.jsx
